requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cssselect==1.2.0
cycler==0.12.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
fonttools==4.53.1
idna==3.10
jieba3k==0.35.1
joblib==1.4.2
kiwisolver==1.4.7
lxml==5.3.0
lxml_html_clean==0.2.2
matplotlib==3.9.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.6
textblob==0.18.0.post0
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.66.5
tzdata==2024.1
urllib3==2.2.3
